uefn_generator/requirements.txt
streamlit>=1.33
anthropic>=0.25
python-dotenv
paramiko
